# api/openapi.yaml
openapi: 3.0.0
info:
  title: Finsolvz Backend API
  description: |
    Comprehensive financial solutions management system with JWT authentication, 
    user management, company management, report type management, and complete report management.
    
    **Authentication Required**: Most endpoints require a valid JWT token in the Authorization header.
    
    **Roles & Authorization**:
    - `SUPER_ADMIN`: Full system access including user registration, role management, and company/user operations
    - `ADMIN`: User management access (get users list)  
    - `CLIENT`: Basic authenticated access to reports and own profile
    
    **Authorization Pattern**: Authorization checks are performed at the controller level, not middleware level.
    Most report endpoints require authentication only, while user management requires specific roles.
    
    **Base URL**: `http://localhost:8787` (development)
  version: 2.0.0
  contact:
    name: Finsolvz Team
    email: support@finsolvz.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8787
    description: Local Development Server
  - url: https://finsolvz-backend.example.com
    description: Production Environment
...
paths:
...
  /api/login:
    post:
      summary: User login
      operationId: login
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/BadRequestError'
  /api/forgot-password:
    post:
      summary: Request password reset
      operationId: forgotPassword
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Password reset email sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "New password has been sent to your email"
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/reset-password:
    post:
      summary: Reset password with token
      operationId: resetPassword
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password successfully reset"
        '400':
          $ref: '#/components/responses/BadRequestError'
...
  /api/company:
    get:
      summary: Get all companies
      operationId: getCompanies
      tags:
        - Company Management
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyResponse'

    post:
      summary: Create new company
      operationId: createCompany
      tags:
        - Company Management
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompanyRequest'
      responses:
        '201':
          description: Company created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Company created successfully"
                  company:
                    $ref: '#/components/schemas/CompanyResponse'
        '409':
          $ref: '#/components/responses/ConflictError'
  /api/company/{id}:
    get:
      summary: Get company by ID
      operationId: getCompanyById
      tags:
        - Company Management
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "60f1b2e5e4b0c7a1d8b9c0d1"
      responses:
        '200':
          description: Company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'

    put:
      summary: Update company (SUPER_ADMIN only)
      operationId: updateCompany
      tags:
        - Company Management
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "60f1b2e5e4b0c7a1d8b9c0d1"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompanyRequest'
      responses:
        '200':
          description: Company updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Success"
                  company:
                    $ref: '#/components/schemas/CompanyResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    delete:
      summary: Delete company (SUPER_ADMIN only)
      operationId: deleteCompany
      tags:
        - Company Management
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            example: "60f1b2e5e4b0c7a1d8b9c0d1"
      responses:
        '200':
          description: Company deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Company deleted successfully"
                  company:
                    $ref: '#/components/schemas/CompanyResponse'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/company/{name}:
    get:
      summary: Get company by name
      operationId: getCompanyByName
      tags:
        - Company Management
      security:
        - BearerAuth: []
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            example: "Acme Corporation"
      responses:
        '200':
          description: Company details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/user/companies:
    get:
      summary: Get current user's companies
      operationId: getUserCompanies
      tags:
        - Company Management
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CompanyResponse'
...
  /api/reports/company/{companyId}:
    get:
      summary: Get reports by company ID
      operationId: getReportsByCompany
      tags:
        - Reports
      security:
        - BearerAuth: []
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: string
            example: "60f1b2e5e4b0c7a1d8b9c0d1"
      responses:
        '200':
          description: List of reports for the company
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReportResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: "admin@finsolvz.com"
        password:
          type: string
          format: password
          example: "admin123"
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: "user@example.com"
    ResetPasswordRequest:
      type: object
      required:
        - token
        - newPassword
      properties:
        token:
          type: string
          example: "reset-token-string"
        newPassword:
          type: string
          format: password
          minLength: 6
          example: "newPassword123!"

    # User Management Schemas
...
    CreateCompanyRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Acme Corporation"
        profilePicture:
          type: string
          format: uri
          example: "https://example.com/logo.png"
        user:
          type: array
          items:
            type: string
          example: ["60f1b2e5e4b0c7a1d8b9c0d1", "60f1b2e5e4b0c7a1d8b9c0d2"]
    UpdateCompanyRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: "Acme Corporation Updated"
        profilePicture:
          type: string
          format: uri
          example: "https://example.com/new-logo.png"
        user:
          type: array
          items:
            type: string
          example: ["60f1b2e5e4b0c7a1d8b9c0d1"]
    CompanyResponse:
      type: object
      properties:
        _id:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d1"
        name:
          type: string
          example: "Acme Corporation"
        profilePicture:
          type: string
          format: uri
          example: "https://example.com/logo.png"
        user:
          type: array
          items:
            $ref: '#/components/schemas/CompanyUserInfo'
        createdAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
    CompanyUserInfo:
      type: object
      properties:
        _id:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d1"
        name:
          type: string
          example: "John Doe"

    # Report Type Schemas
...
    ReportResponse:
      type: object
      properties:
        _id:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d1"
        reportName:
          type: string
          example: "Q4 2024 Financial Report"
        reportType:
          $ref: '#/components/schemas/ReportTypeInfo'
        year:
          type: string
          example: "2024"
          description: "Always returned as string for consistency"
        company:
          $ref: '#/components/schemas/ReportCompanyInfo'
        currency:
          type: string
          example: "USD"
        createdBy:
          $ref: '#/components/schemas/ReportUserInfo'
          description: "User who created the report (populated)"
        userAccess:
          type: array
          items:
            $ref: '#/components/schemas/ReportUserInfo'
          description: "Array of users with access to this report (populated)"
        reportData:
          oneOf:
            - type: array
              items: {}
            - type: object
          example: [{"title": "Revenue", "value": 1000000, "isTotal": false}]
          description: "Report data structure (defaults to empty array if null)"
        createdAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
    ReportTypeInfo:
      type: object
      properties:
        _id:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d1"
        name:
          type: string
          example: "Monthly Financial Report"
    ReportCompanyInfo:
      type: object
      properties:
        _id:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d2"
        name:
          type: string
          example: "Acme Corporation"
        profilePicture:
          type: string
          format: uri
          example: "https://example.com/logo.png"
        createdAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
    ReportUserInfo:
      type: object
      properties:
        _id:
          type: string
          example: "60f1b2e5e4b0c7a1d8b9c0d3"
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
        role:
          type: string
          enum: [SUPER_ADMIN, ADMIN, CLIENT]
          example: "CLIENT"
        createdAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2023-07-15T10:30:00Z"

    # Error Schemas
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Application-specific error code
          example: "INVALID_INPUT"
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request payload"
        details:
          type: string
          description: Additional error details
          example: "Validation failed: email is required"
  responses:
    BadRequestError:
      description: Invalid request payload or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "BAD_REQUEST"
            message: "Invalid request payload or parameters"
            details: "Validation failed: email is required"
    UnauthorizedError:
      description: Authentication required or invalid credentials
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "UNAUTHORIZED"
            message: "Authentication required"
            details: "Invalid or missing JWT token"
    ForbiddenError:
      description: Access denied due to insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "FORBIDDEN"
            message: "Access denied"
            details: "Insufficient permissions for this operation"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
            details: "The requested resource could not be found"
    ConflictError:
      description: Resource conflict or already exists
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "CONFLICT"
            message: "Resource conflict"
            details: "Email already registered"
...